requirements.txt
streamlit
tensorflow-cpu
pandas
numpy
h5py